+++ src/olx/user/UserModel.java
@@ -3,7 +3,7 @@ package olx.user;
 import olx.user.UserConstants.UserStatus;
 
 public class UserModel {
-	int id;
+	long id;
 	String firstName;
 	String lastName;
 	String phone;
@@ -21,7 +21,7 @@ public class UserModel {
 		this.status = status;
 	}
 
-	public UserModel(int id, String firstName, String lastName, String phone, String email, String address,
+	public UserModel(long id, String firstName, String lastName, String phone, String email, String address,
 			boolean admin, UserStatus status) {
 		super();
 		this.id = id;
@@ -46,11 +46,11 @@ public class UserModel {
 		this.password = password;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
